pyproject.toml
[tool.poetry]
name = "aisuite"
version = "0.1.7"
description = "Uniform access layer for LLMs"
authors = ["Andrew Ng, Rohit P"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anthropic = { version = "^0.30.1", optional = true }
boto3 = { version = "^1.34.144", optional = true }
cohere = { version = "^5.12.0", optional = true }
vertexai = { version = "^1.63.0", optional = true }
groq = { version = "^0.9.0", optional = true }
mistralai = { version = "^1.0.3", optional = true }
openai = { version = "^1.35.8", optional = true }
ibm-watsonx-ai = { version = "^1.1.16", optional = true }

# Optional dependencies for different providers
httpx = "~0.27.0"
[tool.poetry.extras]
anthropic = ["anthropic"]
aws = ["boto3"]
azure = []
cohere = ["cohere"]
google = ["vertexai"]
groq = ["groq"]
huggingface = []
mistral = ["mistralai"]
ollama = []
openai = ["openai"]
watsonx = ["ibm-watsonx-ai"]
all = ["anthropic", "aws", "google", "groq", "mistral", "openai", "cohere", "watsonx"]  # To install all providers

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
black = "^24.4.2"
python-dotenv = "^1.0.1"
openai = "^1.35.8"
groq = "^0.9.0"
anthropic = "^0.30.1"
notebook = "^7.2.1"
ollama = "^0.2.1"
mistralai = "^1.0.3"
boto3 = "^1.34.144"
fireworks-ai = "^0.14.0"
chromadb = "^0.5.4"
sentence-transformers = "^3.0.1"
datasets = "^2.20.0"
vertexai = "^1.63.0"
ibm-watsonx-ai = "^1.1.16"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths="tests"
markers = [
    "integration: marks tests as integration tests that interact with external services",
]
